+++ lib/src/constants.dart
@@ -1,16 +1,14 @@
 class Constants {
   // Params names
   static const kApiKey = 'key';
-  static const kUserId = 'userID';
-  static const kAutoTrackPurchases = 'autoTrackPurchases';
+  static const kUserId = 'userId';
   static const kData = 'data';
   static const kProvider = 'provider';
+  static const kDetails = 'details';
+  static const kPurchase = 'purchase';
 
   // MethodChannel methods names
   static const mLaunch = 'launch';
-  static const mLaunchWithKeyCompletion = 'launchWithKeyCompletion';
-  static const mLaunchWithKeyUserId = 'launchWithKeyUserId';
-  static const mLaunchWithKeyAutoTrackPurchasesCompletion =
-      'launchWithKeyAutoTrackPurchasesCompletion';
   static const mAddAttributionData = 'addAttributionData';
+  static const mTrackPurchase = 'trackPurchase';
 }